Fixes a critical bug in the .Current Executor logic (things seem to work now)!
We took a "round trip" to the Swift 2.0 converter and came back.  This code is STILL Swift 1.2 compatible.  But the Swift 2.0 convertor found lots of optimizations that I ported back to swift 1.2
That's gonna make the final Swift 2.0 branch a little easier to deal with.  (Swift 2.0 branch coming soon!)

We are moving the .Error completion value from an NSError associated type to an ErrorType associated type.  And for convenience, ErrorType will 'alias' NSError on Swift 1.2

Coming soon:  Promise that can complete with a 'throwable' block to make Future error handling SUPER EASY in Swift 1.2  (you don't have to catch things yourself!)